circuits: zk-circuits: valid-match-settle: Add settlement tests
