+++ circuits/src/lib.rs
@@ -231,6 +231,13 @@ pub(crate) mod test_helpers {
             timestamp: 0,
         };
 
+        // Randomly permute the orders
+        let (o1, o2) = if rng.gen_bool(0.5) {
+            (o1, o2)
+        } else {
+            (o2, o1)
+        };
+
         // Match orders assuming they are fully capitalized
         let match_res =
             match_orders_with_max_amount(&o1, &o2, o1.amount, o2.amount, price).unwrap();
